+++ packages/components/pages/f-checkout/test/accessibility/axe-accessibility.spec.js
@@ -1,10 +1,7 @@
-import forEach from 'mocha-each';
+import Checkout from '../../test-utils/component-objects/f-checkout.component';
+import CheckoutError from '../../test-utils/component-objects/f-checkout-error.component';
 
-const { getAxeResults } = require('../../../../../../test/utils/axe-helper');
-
-const Checkout = require('../../test-utils/component-objects/f-checkout.component');
-
-let checkout;
+let tests;
 
 describe('Accessibility tests', () => {
     const checkoutInfo = {
@@ -14,111 +11,203 @@ describe('Accessibility tests', () => {
         locale: 'en-GB'
     };
 
-    beforeEach(() => {
-        checkout = new Checkout();
-    });
-
-    it('a11y - should test f-checkout component (delivery) WCAG compliance', () => {
+    it('a11y - should test f-checkout component (delivery) WCAG compliance', async () => {
         // Act
-        checkout.load({
+        await Checkout.load({
             ...checkoutInfo
         });
-        const axeResults = getAxeResults('f-checkout-delivery');
+        const axeResults = await Checkout.getAxeResults('f-checkout-delivery');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);
     });
 
-    it('a11y - should test f-checkout component (collection) WCAG compliance', () => {
+    it('a11y - should test f-checkout component (collection) WCAG compliance', async () => {
         // Act
-        checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             serviceType: 'collection'
         });
-        const axeResults = getAxeResults('f-checkout-collection');
+        const axeResults = await Checkout.getAxeResults('f-checkout-collection');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);
     });
 
-    it('a11y - should test f-checkout component (guest) WCAG compliance', () => {
+    it('a11y - should test f-checkout component (guest) WCAG compliance', async () => {
         // Act
-        checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             isLoggedIn: false
         });
-        const axeResults = getAxeResults('f-checkout-guest');
+        const axeResults = await Checkout.getAxeResults('f-checkout-guest');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);
     });
 
-    it('a11y - should test f-checkout component (error) WCAG compliance', () => {
+    it('a11y - should test f-checkout component (error) WCAG compliance', async () => {
         // Act
-        checkout.load({
+        await CheckoutError.load({
             ...checkoutInfo,
             serviceType: 'invalid-url',
             isLoggedIn: false
         });
-        const axeResults = getAxeResults('f-checkout-error-page');
+        const axeResults = await CheckoutError.getAxeResults('f-checkout-error-page');
 
         // Assert
         expect(axeResults.violations.length).toBe(0);
     });
 
-    forEach([
-        ['collection', 'restaurant-not-taking-orders'],
-        ['collection', 'additional-items-required'],
-        ['collection', 'time-unavailable'],
-        ['delivery', 'restaurant-not-taking-orders'],
-        ['delivery', 'additional-items-required'],
-        ['delivery', 'time-unavailable'],
-        ['dinein', 'restaurant-not-taking-orders'],
-        ['dinein', 'additional-items-required'],
-        ['dinein', 'time-unavailable']
-    ])
-    .it('a11y - Authenticated - "%s" - should have a correct tab order in patch checkout error - "%s"', (serviceType, patchCheckoutError) => {
-        // Act
-        checkout.load({
-            ...checkoutInfo,
-            serviceType,
-            patchCheckoutError
+    tests = [
+        { serviceType: 'collection', patchCheckoutError: 'restaurant-not-taking-orders' },
+        { serviceType: 'collection', patchCheckoutError: 'additional-items-required' },
+        { serviceType: 'collection', patchCheckoutError: 'time-unavailable' },
+        { serviceType: 'delivery', patchCheckoutError: 'restaurant-not-taking-orders' },
+        { serviceType: 'delivery', patchCheckoutError: 'additional-items-required' },
+        { serviceType: 'delivery', patchCheckoutError: 'time-unavailable' },
+        { serviceType: 'dinein', patchCheckoutError: 'restaurant-not-taking-orders' },
+        { serviceType: 'dinein', patchCheckoutError: 'additional-items-required' },
+        { serviceType: 'dinein', patchCheckoutError: 'time-unavailable' }
+    ];
+
+    tests.forEach(({ serviceType, patchCheckoutError }) => {
+        it(`a11y - Authenticated - "${serviceType}" - should have a correct tab order in patch checkout error - "${patchCheckoutError}"`, async () => {
+            // Act
+            await Checkout.load({
+                ...checkoutInfo,
+                serviceType,
+                patchCheckoutError
+            });
+            await Checkout.goToPayment();
+
+            const expectedTabOrder = Promise.all([Checkout.errorMessageRetry, Checkout.closeMessageModal]);
+            const result = Checkout.testTabOrder(expectedTabOrder);
+
+            // Assert
+            await expect(result.actual).toEqual(result.expected);
         });
-        checkout.goToPayment();
-
-        const expectedTabOrder = [checkout.errorMessageRetry, checkout.closeMessageModal];
-        const result = checkout.testTabOrder(expectedTabOrder);
-
-        // Assert
-        expect(result.actual).toEqual(result.expected);
     });
 
-    forEach([
-        ['collection'],
-        ['delivery']
-    ])
-    .it('a11y - Authenticated - "%s" - should have a correct tab order in duplicate order error', serviceType => {
-        // Act
-        checkout.load({
-            ...checkoutInfo,
-            serviceType,
-            placeOrderError: 'duplicate'
+    tests = [
+        'collection',
+        'delivery'
+    ];
+
+    tests.forEach(serviceType => {
+        it(`a11y - Authenticated - "${serviceType}" - should have a correct tab order in duplicate order error`, async () => {
+            // Act
+            await Checkout.load({
+                ...checkoutInfo,
+                serviceType,
+                placeOrderError: 'duplicate'
+            });
+            await Checkout.goToPayment();
+
+            const expectedTabOrder = Promise.all([Checkout.errorMessageRetry, Checkout.errorMessageDupOrderGoToHistory, Checkout.closeMessageModal]);
+            const result = Checkout.testTabOrder(expectedTabOrder);
+
+            // Assert
+            await expect(result.actual).toEqual(result.expected);
         });
-        checkout.goToPayment();
-
-        const expectedTabOrder = [checkout.errorMessageRetry, checkout.errorMessageDupOrderGoToHistory, checkout.closeMessageModal];
-        const result = checkout.testTabOrder(expectedTabOrder);
-
-        // Assert
-        expect(result.actual).toEqual(result.expected);
     });
 
-    forEach([
-        ['restaurant-not-taking-orders'],
-        ['additional-items-required'],
-        ['time-unavailable']
-    ])
-    .it('a11y - Guest - Collection - should have a correct tab order in patch checkout error - "%s"', patchCheckoutError => {
+    tests = [
+        'restaurant-not-taking-orders',
+        'additional-items-required',
+        'time-unavailable'
+    ];
+
+    tests.forEach(patchCheckoutError => {
+        it(`a11y - Guest - Collection - should have a correct tab order in patch checkout error - "${patchCheckoutError}"`, async () => {
+            // Arrange
+            const customerInfo = {
+                firstName: 'Jerry',
+                lastName: 'Jazzman',
+                emailAddress: 'jerry.jazzman@ronniescotts.co.uk',
+                mobileNumber: '+447111111111'
+            };
+
+            Checkout.inputFieldValues = customerInfo;
+
+            // Act
+            await Checkout.load({
+                ...checkoutInfo,
+                serviceType: 'collection',
+                isLoggedIn: false,
+                patchCheckoutError
+            });
+            await Checkout.setFieldValues();
+            await Checkout.goToPayment();
+
+            const expectedTabOrder = Promise.all([Checkout.errorMessageRetry, Checkout.closeMessageModal]);
+            const result = Checkout.testTabOrder(expectedTabOrder);
+
+            // Assert
+            await expect(result.actual).toEqual(result.expected);
+        });
+
+        it(`a11y - Guest - Delivery - should have a correct tab order in patch checkout error - "${patchCheckoutError}"`, async () => {
+            // Arrange
+            const customerInfo = {
+                firstName: 'Jerry',
+                lastName: 'Jazzman',
+                emailAddress: 'jerry.jazzman@ronniescotts.co.uk',
+                mobileNumber: '+447111111111',
+                addressLine1: '1 Bristol Road',
+                addressLocality: 'Bristol',
+                addressPostcode: 'BS1 1AA'
+            };
+
+            Checkout.inputFieldValues = customerInfo;
+
+            // Act
+            await Checkout.load({
+                ...checkoutInfo,
+                isLoggedIn: false,
+                patchCheckoutError
+            });
+            await Checkout.setFieldValues();
+            await Checkout.goToPayment();
+
+            const expectedTabOrder = Promise.all([Checkout.errorMessageRetry, Checkout.closeMessageModal]);
+            const result = Checkout.testTabOrder(expectedTabOrder);
+
+            // Assert
+            await expect(result.actual).toEqual(result.expected);
+        });
+
+        it(`a11y - Guest - Dine In - should have a correct tab order in patch checkout error - "${patchCheckoutError}"`, async () => {
+            // Arrange
+            const customerInfo = {
+                firstName: 'Jerry',
+                lastName: 'Jazzman',
+                emailAddress: 'jerry.jazzman@ronniescotts.co.uk',
+                mobileNumber: '+447111111111',
+                tableIdentifier: '1'
+            };
+
+            Checkout.inputFieldValues = customerInfo;
+
+            // Act
+            await Checkout.load({
+                ...checkoutInfo,
+                serviceType: 'dinein',
+                isLoggedIn: false,
+                patchCheckoutError
+            });
+            await Checkout.setFieldValues();
+            await Checkout.goToPayment();
+
+            const expectedTabOrder = Promise.all([Checkout.errorMessageRetry, Checkout.closeMessageModal]);
+            const result = Checkout.testTabOrder(expectedTabOrder);
+
+            // Assert
+            await expect(result.actual).toEqual(result.expected);
+        });
+    });
+
+    it('a11y - Guest - Collection - should have a correct tab order in duplicate order error', async () => {
         // Arrange
         const customerInfo = {
             firstName: 'Jerry',
@@ -127,31 +216,26 @@ describe('Accessibility tests', () => {
             mobileNumber: '+447111111111'
         };
 
-        checkout.inputFieldValues = customerInfo;
+        Checkout.inputFieldValues = customerInfo;
 
         // Act
-        checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             serviceType: 'collection',
             isLoggedIn: false,
-            patchCheckoutError
+            placeOrderError: 'duplicate'
         });
-        checkout.setFieldValues();
-        checkout.goToPayment();
+        await Checkout.setFieldValues();
+        await Checkout.goToPayment();
 
-        const expectedTabOrder = [checkout.errorMessageRetry, checkout.closeMessageModal];
-        const result = checkout.testTabOrder(expectedTabOrder);
+        const expectedTabOrder = Promise.all([Checkout.errorMessageRetry, Checkout.errorMessageDupOrderGoToHistory, Checkout.closeMessageModal]);
+        const result = Checkout.testTabOrder(expectedTabOrder);
 
         // Assert
-        expect(result.actual).toEqual(result.expected);
+        await expect(result.actual).toEqual(result.expected);
     });
 
-    forEach([
-        ['restaurant-not-taking-orders'],
-        ['additional-items-required'],
-        ['time-unavailable']
-    ])
-    .it('a11y - Guest - Delivery - should have a correct tab order in patch checkout error - "%s"', patchCheckoutError => {
+    it('a11y - Guest - Delivery - should have a correct tab order in duplicate order error', async () => {
         // Arrange
         const customerInfo = {
             firstName: 'Jerry',
@@ -163,30 +247,25 @@ describe('Accessibility tests', () => {
             addressPostcode: 'BS1 1AA'
         };
 
-        checkout.inputFieldValues = customerInfo;
+        Checkout.inputFieldValues = customerInfo;
 
         // Act
-        checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             isLoggedIn: false,
-            patchCheckoutError
+            placeOrderError: 'duplicate'
         });
-        checkout.setFieldValues();
-        checkout.goToPayment();
+        await Checkout.setFieldValues();
+        await Checkout.goToPayment();
 
-        const expectedTabOrder = [checkout.errorMessageRetry, checkout.closeMessageModal];
-        const result = checkout.testTabOrder(expectedTabOrder);
+        const expectedTabOrder = Promise.all([Checkout.errorMessageRetry, Checkout.errorMessageDupOrderGoToHistory, Checkout.closeMessageModal]);
+        const result = Checkout.testTabOrder(expectedTabOrder);
 
         // Assert
-        expect(result.actual).toEqual(result.expected);
+        await expect(result.actual).toEqual(result.expected);
     });
 
-    forEach([
-        ['restaurant-not-taking-orders'],
-        ['additional-items-required'],
-        ['time-unavailable']
-    ])
-    .it('a11y - Guest - Dine In - should have a correct tab order in patch checkout error - "%s"', patchCheckoutError => {
+    it('a11y - Guest - Dine In - should have a correct tab order in duplicate order error', async () => {
         // Arrange
         const customerInfo = {
             firstName: 'Jerry',
@@ -196,109 +275,22 @@ describe('Accessibility tests', () => {
             tableIdentifier: '1'
         };
 
-        checkout.inputFieldValues = customerInfo;
+        Checkout.inputFieldValues = customerInfo;
 
         // Act
-        checkout.load({
-            ...checkoutInfo,
-            serviceType: 'dinein',
-            isLoggedIn: false,
-            patchCheckoutError
-        });
-        checkout.setFieldValues();
-        checkout.goToPayment();
-
-        const expectedTabOrder = [checkout.errorMessageRetry, checkout.closeMessageModal];
-        const result = checkout.testTabOrder(expectedTabOrder);
-
-        // Assert
-        expect(result.actual).toEqual(result.expected);
-    });
-
-    it('a11y - Guest - Collection - should have a correct tab order in duplicate order error', () => {
-        // Arrange
-        const customerInfo = {
-            firstName: 'Jerry',
-            lastName: 'Jazzman',
-            emailAddress: 'jerry.jazzman@ronniescotts.co.uk',
-            mobileNumber: '+447111111111'
-        };
-
-        checkout.inputFieldValues = customerInfo;
-
-        // Act
-        checkout.load({
-            ...checkoutInfo,
-            serviceType: 'collection',
-            isLoggedIn: false,
-            placeOrderError: 'duplicate'
-        });
-        checkout.setFieldValues();
-        checkout.goToPayment();
-
-        const expectedTabOrder = [checkout.errorMessageRetry, checkout.errorMessageDupOrderGoToHistory, checkout.closeMessageModal];
-        const result = checkout.testTabOrder(expectedTabOrder);
-
-        // Assert
-        expect(result.actual).toEqual(result.expected);
-    });
-
-    it('a11y - Guest - Delivery - should have a correct tab order in duplicate order error', () => {
-        // Arrange
-        const customerInfo = {
-            firstName: 'Jerry',
-            lastName: 'Jazzman',
-            emailAddress: 'jerry.jazzman@ronniescotts.co.uk',
-            mobileNumber: '+447111111111',
-            addressLine1: '1 Bristol Road',
-            addressLocality: 'Bristol',
-            addressPostcode: 'BS1 1AA'
-        };
-
-        checkout.inputFieldValues = customerInfo;
-
-        // Act
-        checkout.load({
-            ...checkoutInfo,
-            isLoggedIn: false,
-            placeOrderError: 'duplicate'
-        });
-        checkout.setFieldValues();
-        checkout.goToPayment();
-
-        const expectedTabOrder = [checkout.errorMessageRetry, checkout.errorMessageDupOrderGoToHistory, checkout.closeMessageModal];
-        const result = checkout.testTabOrder(expectedTabOrder);
-
-        // Assert
-        expect(result.actual).toEqual(result.expected);
-    });
-
-    it('a11y - Guest - Dine In - should have a correct tab order in duplicate order error', () => {
-        // Arrange
-        const customerInfo = {
-            firstName: 'Jerry',
-            lastName: 'Jazzman',
-            emailAddress: 'jerry.jazzman@ronniescotts.co.uk',
-            mobileNumber: '+447111111111',
-            tableIdentifier: '1'
-        };
-
-        checkout.inputFieldValues = customerInfo;
-
-        // Act
-        checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             serviceType: 'dinein',
             isLoggedIn: false,
             placeOrderError: 'duplicate'
         });
-        checkout.setFieldValues();
-        checkout.goToPayment();
+        await Checkout.setFieldValues();
+        await Checkout.goToPayment();
 
-        const expectedTabOrder = [checkout.errorMessageRetry, checkout.errorMessageDupOrderGoToHistory, checkout.closeMessageModal];
-        const result = checkout.testTabOrder(expectedTabOrder);
+        const expectedTabOrder = Promise.all([Checkout.errorMessageRetry, Checkout.errorMessageDupOrderGoToHistory, Checkout.closeMessageModal]);
+        const result = Checkout.testTabOrder(expectedTabOrder);
 
         // Assert
-        expect(result.actual).toEqual(result.expected);
+        await expect(result.actual).toEqual(result.expected);
     });
 });