+++ packages/components/pages/f-checkout/test/visual/f-checkout-authenticated.visual.desktop.spec.js
@@ -1,10 +1,8 @@
-import forEach from 'mocha-each';
-
-const Checkout = require('../../test-utils/component-objects/f-checkout.component');
-
-let checkout;
+import Checkout from '../../test-utils/component-objects/f-checkout.component';
+import CheckoutAgeVerification from '../../test-utils/component-objects/f-checkout-age-verification';
 
 const illegalMobileNumber = '123';
+let tests;
 
 describe('f-checkout - Collection - Authenticated - Desktop Visual Tests', () => {
     const checkoutInfo = {
@@ -14,26 +12,25 @@ describe('f-checkout - Collection - Authenticated - Desktop Visual Tests', () =>
         locale: 'en-GB'
     };
 
-    beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
-        // Act
-        await checkout.load({
-            ...checkoutInfo,
-            placeOrderError: 'duplicate'
-        });
-    });
-
     it('should display the component base state.', async () => {
+        // Act
+        await Checkout.load({
+            ...checkoutInfo
+        });
+
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - Base State', 'desktop');
     });
 
     it('should display the mandatory error messages.', async () => {
         // Act
-        await checkout.clearBlurField('mobileNumber');
-        await checkout.goToPayment();
+        await Checkout.load({
+            ...checkoutInfo
+        });
+
+        // Act
+        await Checkout.clearBlurField('mobileNumber');
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - Mandatory Errors State', 'desktop');
@@ -41,11 +38,11 @@ describe('f-checkout - Collection - Authenticated - Desktop Visual Tests', () =>
 
     it('should display the "Something went wrong" error.', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             patchCheckoutError: 'SERVER'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
         await browser.pause(500);
 
         // Assert
@@ -54,11 +51,11 @@ describe('f-checkout - Collection - Authenticated - Desktop Visual Tests', () =>
 
     it('should display the "Restaurant not taking orders" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             patchCheckoutError: 'restaurant-not-taking-orders'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - "Restaurant not taking orders" Error Modal', 'desktop');
@@ -66,11 +63,11 @@ describe('f-checkout - Collection - Authenticated - Desktop Visual Tests', () =>
 
     it('should display the "Additional Items Required" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             patchCheckoutError: 'additional-items-required'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - "Additional Items Required" Error Modal', 'desktop');
@@ -78,9 +75,14 @@ describe('f-checkout - Collection - Authenticated - Desktop Visual Tests', () =>
 
     it('should display the illegal mobile number error message', async () => {
         // Act
-        await checkout.clearBlurField('mobileNumber');
-        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        await checkout.goToPayment();
+        await Checkout.load({
+            ...checkoutInfo
+        });
+
+        // Act
+        await Checkout.clearBlurField('mobileNumber');
+        await Checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - Illegal Mobile Number Error State', 'desktop');
@@ -88,11 +90,11 @@ describe('f-checkout - Collection - Authenticated - Desktop Visual Tests', () =>
 
     it('should display the "Duplicate Order Warning" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Collection - Authenticated - "Duplicate Order Warning" Modal', 'desktop');
@@ -101,11 +103,8 @@ describe('f-checkout - Collection - Authenticated - Desktop Visual Tests', () =>
 
 describe('f-checkout - Collection - Authenticated - isAsapAvailable: false Desktop Visual Tests', () => {
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             serviceType: 'collection',
             isLoggedIn: true,
             isAsapAvailable: false,
@@ -128,11 +127,8 @@ describe('f-checkout - Delivery - Authenticated - Desktop Visual Tests', () => {
     };
 
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo
         });
     });
@@ -142,14 +138,23 @@ describe('f-checkout - Delivery - Authenticated - Desktop Visual Tests', () => {
         await browser.percyScreenshot('f-checkout - Delivery - Authenticated - Base State', 'desktop');
     });
 
-    forEach(['addressLine1', 'addressLocality', 'mobileNumber', 'addressPostcode'])
-    .it('should display the mandatory error messages', async field => {
-        // Act
-        await checkout.clearBlurField(field);
-        await checkout.goToPayment();
 
-        // Assert
-        await browser.percyScreenshot(`f-checkout - Delivery - Authenticated - Manadatory Error - ${field}`, 'desktop');
+    tests = [
+        'addressLine1',
+        'addressLocality',
+        'mobileNumber',
+        'addressPostcode'
+    ];
+
+    tests.forEach(field => {
+        it('should display the mandatory error messages', async () => {
+            // Act
+            await Checkout.clearBlurField(field);
+            await Checkout.goToPayment();
+
+            // Assert
+            await browser.percyScreenshot(`f-checkout - Delivery - Authenticated - Manadatory Error - ${field}`, 'desktop');
+        });
     });
 
     it('should display the illegal postcode error message', async () => {
@@ -157,9 +162,9 @@ describe('f-checkout - Delivery - Authenticated - Desktop Visual Tests', () => {
         const addressPostcode = 'TEST1A';
 
         // Act
-        await checkout.clearBlurField('addressPostcode');
-        await checkout.setFieldValue('addressPostcode', addressPostcode);
-        await checkout.goToPayment();
+        await Checkout.clearBlurField('addressPostcode');
+        await Checkout.setFieldValue('addressPostcode', addressPostcode);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Authenticated - Illegal Postcode Error State', 'desktop');
@@ -167,9 +172,9 @@ describe('f-checkout - Delivery - Authenticated - Desktop Visual Tests', () => {
 
     it('should display the illegal mobile number error message', async () => {
         // Act
-        await checkout.clearBlurField('mobileNumber');
-        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        await checkout.goToPayment();
+        await Checkout.clearBlurField('mobileNumber');
+        await Checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Authenticated - Illegal Mobile Number Error State', 'desktop');
@@ -177,11 +182,11 @@ describe('f-checkout - Delivery - Authenticated - Desktop Visual Tests', () => {
 
     it('should display the "Duplicate Order Warning" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Delivery - Authenticated - "Duplicate Order Warning" Modal', 'desktop');
@@ -189,7 +194,7 @@ describe('f-checkout - Delivery - Authenticated - Desktop Visual Tests', () => {
 
     it('should display the two notes fields if there is two noteTypes.', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             noteType: 'get-notes-config-split'
         });
@@ -201,11 +206,8 @@ describe('f-checkout - Delivery - Authenticated - Desktop Visual Tests', () => {
 
 describe('f-checkout - Delivery - Authenticated - isAsapAvailable: false Desktop Visual Tests', () => {
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             serviceType: 'delivery',
             isLoggedIn: true,
             isAsapAvailable: false,
@@ -228,11 +230,8 @@ describe('f-checkout - Dine In - Authenticated - Desktop Visual Tests', () => {
     };
 
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo
         });
     });
@@ -244,8 +243,8 @@ describe('f-checkout - Dine In - Authenticated - Desktop Visual Tests', () => {
 
     it('should display the mandatory error messages', async () => {
         // Act
-        ['tableIdentifier', 'mobileNumber'].forEach(async field => checkout.clearBlurField(field));
-        await checkout.goToPayment();
+        ['tableIdentifier', 'mobileNumber'].forEach(async field => Checkout.clearBlurField(field));
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Dine In - Authenticated - Manadatory Errors', 'desktop');
@@ -253,9 +252,9 @@ describe('f-checkout - Dine In - Authenticated - Desktop Visual Tests', () => {
 
     it('should display the illegal mobile number error message', async () => {
         // Act
-        await checkout.clearBlurField('mobileNumber');
-        await checkout.setFieldValue('mobileNumber', illegalMobileNumber);
-        await checkout.goToPayment();
+        await Checkout.clearBlurField('mobileNumber');
+        await Checkout.setFieldValue('mobileNumber', illegalMobileNumber);
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Dine In - Authenticated - Illegal Mobile Number Error State', 'desktop');
@@ -263,11 +262,11 @@ describe('f-checkout - Dine In - Authenticated - Desktop Visual Tests', () => {
 
     it('should display the "Duplicate Order Warning" modal', async () => {
         // Act
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             placeOrderError: 'duplicate'
         });
-        await checkout.goToPayment();
+        await Checkout.goToPayment();
 
         // Assert
         await browser.percyScreenshot('f-checkout - Dine in - Authenticated - "Duplicate Order Warning" Modal', 'desktop');
@@ -275,7 +274,7 @@ describe('f-checkout - Dine In - Authenticated - Desktop Visual Tests', () => {
 
     it('should display the two notes fields if there is two noteTypes.', async () => {
         // Arrange
-        await checkout.load({
+        await Checkout.load({
             ...checkoutInfo,
             noteType: 'get-notes-config-split'
         });
@@ -287,11 +286,8 @@ describe('f-checkout - Dine In - Authenticated - Desktop Visual Tests', () => {
 
 describe('f-checkout - Delivery - AU Tenant - visibile state field - Desktop Visual Tests', () => {
     beforeEach(async () => {
-        // Arrange
-        checkout = new Checkout();
-
         // Act
-        await checkout.load({
+        await Checkout.load({
             serviceType: 'delivery',
             isLoggedIn: true,
             isAsapAvailable: false,
@@ -306,12 +302,9 @@ describe('f-checkout - Delivery - AU Tenant - visibile state field - Desktop Vis
 });
 
 describe('f-checkout - Delivery - AU Tenant - age verification page - Desktop Visual Tests', () => {
-    beforeEach(() => {
-        // Arrange
-        checkout = new Checkout();
-
+    beforeEach(async () => {
         // Act
-        checkout.load({
+        await CheckoutAgeVerification.load({
             serviceType: 'delivery',
             isLoggedIn: true,
             isAsapAvailable: true,